Fix type conversion warnings with OpenSSL
with PISTACHE_USE_SSL

Fix type conversion (size_t -> int) warnings occurring when using
-Wconversion flag of gcc and PISTACHE_USE_SSL is defined

OpenSSL 1.1.1 has new functions SSL_read_ex / SSL_write_ex which allow
size_t parameter, but 1.0.2 has only SSL_read / SSL_write with int
type parameter

+++ src/common/transport.cc
@@ -125,7 +125,7 @@ void Transport::handleIncoming(const std::shared_ptr<Peer> &peer) {
 #ifdef PISTACHE_USE_SSL
     if (peer->ssl() != NULL) {
       bytes = SSL_read((SSL *)peer->ssl(), buffer + totalBytes,
-                       Const::MaxBuffer - totalBytes);
+                       static_cast<int>(Const::MaxBuffer - totalBytes));
     } else {
 #endif /* PISTACHE_USE_SSL */
       bytes = recv(fd, buffer + totalBytes, Const::MaxBuffer - totalBytes, 0);
@@ -228,7 +228,7 @@ void Transport::asyncWriteImpl(Fd fd) {
 
         if (it_->second->ssl() != NULL) {
           auto ssl_ = static_cast<SSL *>(it_->second->ssl());
-          bytesWritten = SSL_write(ssl_, ptr, len);
+          bytesWritten = SSL_write(ssl_, ptr, static_cast<int>(len));
         } else {
 #endif /* PISTACHE_USE_SSL */
           bytesWritten = ::send(fd, ptr, len, flags);
